Refactored data types
