+++ src/main/java/com/hierynomus/msdtyp/AccessMask.java
@@ -39,10 +39,7 @@ public enum AccessMask implements EnumWithValue<AccessMask> {
     ADS_RIGHT_DS_DELETE_CHILD(0X00000002L),
     ADS_RIGHT_DS_READ_PROP(0x00000010L),
     ADS_RIGHT_DS_WRITE_PROP(0x00000020L),
-    ADS_RIGHT_DS_SELF(0x00000008)
-
-    ;
-
+    ADS_RIGHT_DS_SELF(0x00000008L);
 
     private long value;
